wip: sqlite writer

+++ src/utils/list.h
@@ -8,7 +8,7 @@
 
 /**
  * @file list.h 
- * @author Jouni Malinen
+ * @authora Jouni Malinen, Alexandru Mereacre
  * @brief Doubly-linked list.
  */
 
@@ -57,10 +57,10 @@ static inline int dl_list_empty(struct dl_list *list)
 	return list->next == list;
 }
 
-static inline unsigned int dl_list_len(struct dl_list *list)
+static inline ssize_t dl_list_len(struct dl_list *list)
 {
 	struct dl_list *item;
-	int count = 0;
+	ssize_t count = 0;
 	for (item = list->next; item != list; item = item->next)
 		count++;
 	return count;
